+++ sevici_api/data_manager/serializers.py
@@ -1,0 +1,8 @@
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+# Serializers define the API representation.
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = User
+        fields = ['url', 'username', 'email', 'is_staff']
